+++ src/arch/Threads/Threads_Win32.h
@@ -20,7 +20,7 @@ public:
 	void Halt( bool Kill );
 	void Resume();
 	uint64_t GetThreadId() const;
-	int Wait();
+	intptr_t Wait();
 };
 
 HANDLE Win32ThreadIdToHandle( uint64_t iID );
